Reorganize sampler tests
